EODReminder.js: removed checkHoliday.js and combined its source code to EODReminder.js
